(자동차 경주) 1단계, 2단계 - 피드백 반영 커밋
