+++ src/main/java/racingcar/step2/CalculatorUtil.java
@@ -1,4 +1,4 @@
-package racingcar_2step;
+package racingcar.step2;
 
 import java.util.Arrays;
 import java.util.List;
@@ -7,6 +7,11 @@ import java.util.Stack;
 public class CalculatorUtil {
 
     public static List<String> list = Arrays.asList("+", "-", "*", "/");
+
+    // private 생성자 추가
+    private CalculatorUtil() {
+        throw new AssertionError();
+    }
 
     public static String[] blankSplit(String input) {
         return input.split(" ");
