+++ src/main/java/racingcar/step2/operation/Operation.java
@@ -1,7 +1,6 @@
-package racingcar_2step.operation;
+package racingcar.step2.operation;
 
 public interface Operation {
-
     String operate();
 
     String getSign();
