+++ src/test/java/racingcar/step1/SetTest.java
@@ -1,4 +1,4 @@
-package racingcar_1step;
+package racingcar.step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
