+++ src/test/java/racingcar/step1/StringClassTest.java
@@ -1,4 +1,4 @@
-package racingcar_1step;
+package racingcar.step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
